Move Functionality in Hbase
